uldso: fix ARM relocation processing
Unlike M68K and RISCV architecture ELF format binaries ARM platforms use
the ELF "rel" section and not the "rela" section for their relocations.
It is similar but not identical and needs to be processed slightly
differently.

For the most part the procedure followed is the same, just the section
identifiers and relocation data structure is different. With a bit of
C-preprocessor abstracting the code is mostly the same.

Signed-off-by: Greg Ungerer <gerg@kernel.org>