uldso: add support for xtensa architecture
Add support for running on the xtensa architecture. It supports a number
of nommu processor variants and platforms. It has had flat format binary
support for a long time now and ELF-FDPIC is available for it.

Given the support for ELF-FDPIC it is relatively easy to port in and use
uldso to run PIE style ELF binaries on xtensa.

Signed-off-by: Greg Ungerer <gerg@kernel.org>