uldso: implement an alternate argument layout for ARm
Implement some logic in the exit to executable path that converts the
stack argc/argv layout from the ELF-FDPIC style (which is also the
normal VM linux layout) to the uclinux argc/argv layout. With this in
place you do not need to patch uClibc to use ELF-FDPIC layout instead of
uclinuc layout (which is the default when compiling for noMMU).

Signed-off-by: Greg Ungerer <gerg@kernel.org>